suppress (stupid) warnings

+++ include/libterm/output.h
@@ -19,7 +19,7 @@ typedef struct {
     uint8_t b;
 } Libterm_Color_24;
 
-#define LIBTERM_CREATE_COLOR(hex) (Libterm_Color_24){ .r = ((uint32_t)hex >> 16) & 0xFF, .g = ((uint32_t)hex >> 8) & 0xFF, .b = hex & 0xFF } 
+#define LIBTERM_CREATE_COLOR(hex) (Libterm_Color_24){ .r = (uint8_t)(((uint32_t)hex >> 16) & 0xFF), .g = (uint8_t)(((uint32_t)hex >> 8) & 0xFF), .b = (uint8_t)(hex & 0xFF) } 
 
 LIBTERM_EXTERN Libterm_Result libterm_set_foreground_color(Libterm_Color_24 color);
 LIBTERM_EXTERN Libterm_Result libterm_set_background_color(Libterm_Color_24 color);
